Moved some constants to constants.py
