+++ tasks.py
@@ -11,7 +11,15 @@ from pathlib import Path
 import docker
 import git
 import requests
-import tests
+import tests.constants as constants
+from tests.test import (
+    run_aws_stage,
+    run_az_stage,
+    run_cli,
+    run_security,
+    run_terraform,
+    version_commands,
+)
 from invoke import task
 from jinja2 import Environment, FileSystemLoader
 from yaml import YAMLError, dump, safe_load
@@ -173,43 +181,8 @@ VERSION = CONFIG["version"]
 CWD = Path(".").absolute()
 TESTS_PATH = CWD.joinpath("tests")
 
-LOG_FORMAT = json.dumps(
-    {
-        "timestamp": "%(asctime)s",
-        "namespace": "%(name)s",
-        "loglevel": "%(levelname)s",
-        "message": "%(message)s",
-    }
-)
-basicConfig(level="INFO", format=LOG_FORMAT)
 LOG = getLogger("easy_infra")
 
-# git
-REPO = git.Repo(CWD)
-COMMIT_HASH = REPO.head.object.hexsha
-
-# Docker
-CLIENT = docker.from_env()
-IMAGE = "seiso/easy_infra"
-TARGETS = {
-    "minimal": {},
-    "aws": {},
-    "az": {},
-    "final": {},
-}
-for target in TARGETS:
-    if target == "final":
-        TARGETS[target]["tags"] = [
-            IMAGE + ":" + COMMIT_HASH,
-            IMAGE + ":" + VERSION,
-            IMAGE + ":latest",
-        ]
-    else:
-        TARGETS[target]["tags"] = [
-            IMAGE + ":" + COMMIT_HASH + "-" + target,
-            IMAGE + ":" + VERSION + "-" + target,
-            IMAGE + ":" + "latest" + "-" + target,
-        ]
 
 # easy_infra
 APT_PACKAGES = {"ansible", "azure-cli"}
@@ -245,7 +218,7 @@ def update(c):  # pylint: disable=unused-argument
     image = "python:3.9"
     working_dir = "/usr/src/app/"
     volumes = {CWD: {"bind": working_dir, "mode": "rw"}}
-    CLIENT.images.pull(repository=image)
+    constants.CLIENT.images.pull(repository=image)
     command = '/bin/bash -c "python3 -m pip install --upgrade pipenv &>/dev/null && pipenv update"'
     opinionated_docker_run(
         image=image,
@@ -263,7 +236,7 @@ def lint(c):  # pylint: disable=unused-argument
     image = "projectatomic/dockerfile-lint"
     working_dir = "/root/"
     volumes = {CWD: {"bind": working_dir, "mode": "ro"}}
-    CLIENT.images.pull(repository=image)
+    constants.CLIENT.images.pull(repository=image)
     opinionated_docker_run(
         image=image,
         volumes=volumes,
@@ -277,18 +250,18 @@ def build(c):  # pylint: disable=unused-argument
     """Build easy_infra"""
     render_jinja2(template_file=JINJA2_FILE, config=CONFIG, output_file=OUTPUT_FILE)
 
-    buildargs = {"VERSION": VERSION, "COMMIT_HASH": COMMIT_HASH}
-    for command in CONFIG["commands"]:
-        if "version" in CONFIG["commands"][command]:
+    buildargs = {"VERSION": constants.VERSION, "COMMIT_HASH": constants.COMMIT_HASH}
+    for command in constants.CONFIG["commands"]:
+        if "version" in constants.CONFIG["commands"][command]:
             # Normalize the build args
             arg = command.upper().replace("-", "_") + "_VERSION"
-            buildargs[arg] = CONFIG["commands"][command]["version"]
+            buildargs[arg] = constants.CONFIG["commands"][command]["version"]
 
     # pylint: disable=redefined-outer-name
-    for target in TARGETS:
-        for tag in TARGETS[target]["tags"]:
+    for target in constants.TARGETS:
+        for tag in constants.TARGETS[target]["tags"]:
             LOG.info("Building %s...", tag)
-            CLIENT.images.build(
+            constants.CLIENT.images.build(
                 path=str(CWD), target=target, rm=True, tag=tag, buildargs=buildargs
             )
 
@@ -300,27 +273,27 @@ def test(c):  # pylint: disable=unused-argument
     default_volumes = {TESTS_PATH: {"bind": default_working_dir, "mode": "ro"}}
 
     # pylint: disable=redefined-outer-name
-    for target in TARGETS:
+    for target in constants.TARGETS:
         # Only test using the last tag for each target
-        image = TARGETS[target]["tags"][-1]
+        image = constants.TARGETS[target]["tags"][-1]
 
         LOG.info("Testing %s...", image)
         if target == "minimal":
-            tests.run_terraform(image=image)
-            tests.run_security(image=image)
+            run_terraform(image=image)
+            run_security(image=image)
         elif target == "az":
-            tests.run_az_stage(image=image)
-            tests.run_security(image=image)
+            run_az_stage(image=image)
+            run_security(image=image)
         elif target == "aws":
-            tests.run_aws_stage(image=image)
-            tests.run_security(image=image)
+            run_aws_stage(image=image)
+            run_security(image=image)
         elif target == "final":
-            tests.version_commands(
+            version_commands(
                 image=image, volumes=default_volumes, working_dir=default_working_dir
             )
-            tests.run_terraform(image=image)
-            tests.run_cli(image=image)
-            tests.run_security(image=image)
+            run_terraform(image=image)
+            run_cli(image=image)
+            run_security(image=image)
         else:
             LOG.error("Untested stage of %s", target)
 
@@ -329,11 +302,11 @@ def test(c):  # pylint: disable=unused-argument
 def publish(c):  # pylint: disable=unused-argument
     """Publish easy_infra"""
     # pylint: disable=redefined-outer-name
-    for target in TARGETS:
-        for tag in TARGETS[target]["tags"]:
+    for target in constants.TARGETS:
+        for tag in constants.TARGETS[target]["tags"]:
             repository = tag
             LOG.info("Pushing %s to docker hub...", repository)
-            CLIENT.images.push(repository=repository)
+            constants.CLIENT.images.push(repository=repository)
     LOG.info("Done publishing easy_infra Docker images")
